fix:code optimization done
