+++ client/src/components/admin/Sidebar/menulist/NavGroup.jsx
@@ -1,15 +1,11 @@
-
-// material-ui
+// library
 import { Divider, List, Typography } from '@mui/material';
 
-// project imports
+// components
 import NavItem from './NavItem';
 import NavCollapse from './NavCollapse';
 
-// ==============================|| SIDEBAR MENU LIST GROUP ||============================== //
-
 const NavGroup = ({ item }) => {
-  // menu list collapse & items
   const items = item.children?.map((menu) => {
     switch (menu.type) {
       case 'collapse':
@@ -66,7 +62,6 @@ const NavGroup = ({ item }) => {
         {items}
       </List>
 
-      {/* group divider */}
       <Divider sx={{ mt: 0.25, mb: 1.25 }} />
     </>
   );
