+++ client/src/components/admin/subcomponents/PublishCards.jsx
@@ -1,14 +1,18 @@
 import * as React from 'react';
-import Box from '@mui/material/Box';
-import Card from '@mui/material/Card';
-import CardContent from '@mui/material/CardContent';
-import InputLabel from '@mui/material/InputLabel';
-import MenuItem from '@mui/material/MenuItem';
-import FormControl from '@mui/material/FormControl';
-import Select from '@mui/material/Select';
-import FormGroup from '@mui/material/FormGroup';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import Checkbox from '@mui/material/Checkbox';
+
+// libraries
+import {
+  Box,
+  Card,
+  CardContent,
+  InputLabel,
+  MenuItem,
+  FormControl,
+  Select,
+  FormGroup,
+  FormControlLabel,
+  Checkbox,
+} from '@mui/material';
 import { styled } from '@mui/system';
 
 const Title = styled('span')(() => ({
