+++ client/src/pages/admin/post/add-new.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import Head from 'next/head';
+
+// components
 import AddNew from '../../../screens/admin/AddNew';
 import Admin from '../../../screens/admin/Admin';
